+++ src/internal.h
@@ -540,6 +540,7 @@ struct _GLFWwindow
     GLFWbool            focusOnShow;
     GLFWbool            mousePassthrough;
     GLFWbool            shouldClose;
+    GLFWbool            tiled;
     void*               userPointer;
     GLFWbool            doublebuffer;
     GLFWvidmode         videoMode;